add e2e tests for pass through filter chain for peer authN
