+++ pkg/test/framework/components/echo/common/config.go
@@ -75,6 +75,18 @@ func FillInDefaults(ctx resource.Context, defaultDomain string, c *echo.Config) 
 			portGen.Instance.SetUsed(p.InstancePort)
 		}
 	}
+	for _, p := range c.WorkloadOnlyPorts {
+		if p > 0 {
+			if portGen.Instance.IsUsed(p) {
+				return fmt.Errorf("failed configuring workload only port %d: port already used", p)
+			}
+			portGen.Instance.SetUsed(p)
+			if portGen.Service.IsUsed(p) {
+				return fmt.Errorf("failed configuring workload only port %d: port already used", p)
+			}
+			portGen.Service.SetUsed(p)
+		}
+	}
 
 	// Second pass: try to make unassigned instance ports match service port.
 	for i, p := range c.Ports {
