+++ pkg/test/framework/components/echo/config.go
@@ -58,6 +58,10 @@ type Config struct {
 	// on the implementation.
 	Ports []Port
 
+	// WorkloadOnlyPorts for ports only defined in the workload but not in the k8s service.
+	// This is used to test the inbound pass-through filter chain.
+	WorkloadOnlyPorts []int
+
 	// Annotations provides metadata hints for deployment of the instance.
 	Annotations Annotations
