+++ tests/integration/security/testdata/pass-through-filter-chain.yaml.tmpl
@@ -1,0 +1,73 @@
+# The following policy disables mTLS for workload a.
+
+apiVersion: "security.istio.io/v1beta1"
+kind: "PeerAuthentication"
+metadata:
+  name: "a"
+  namespace: "{{ .Namespace }}"
+spec:
+  selector:
+    matchLabels:
+      app: a
+  mtls:
+    mode: DISABLE
+---
+
+# The following policies disable mTLS for workload b and only allow request to port 8085.
+
+apiVersion: "security.istio.io/v1beta1"
+kind: "PeerAuthentication"
+metadata:
+  name: "b"
+  namespace: "{{ .Namespace }}"
+spec:
+  selector:
+    matchLabels:
+      app: b
+  mtls:
+    mode: DISABLE
+---
+apiVersion: "security.istio.io/v1beta1"
+kind: AuthorizationPolicy
+metadata:
+  name: "b"
+  namespace: "{{ .Namespace }}"
+spec:
+  selector:
+    matchLabels:
+      app: b
+  rules:
+  - to:
+    - operation:
+        ports: ["8085"]
+---
+
+# The following policy enables mTLS (Strict) for workload c.
+
+apiVersion: "security.istio.io/v1beta1"
+kind: "PeerAuthentication"
+metadata:
+  name: "c"
+  namespace: "{{ .Namespace }}"
+spec:
+  selector:
+    matchLabels:
+      app: c
+  mtls:
+    mode: STRICT
+---
+
+# The following policy enables mTLS (Permissive) for workload d.
+
+apiVersion: "security.istio.io/v1beta1"
+kind: "PeerAuthentication"
+metadata:
+  name: "d"
+  namespace: "{{ .Namespace }}"
+spec:
+  selector:
+    matchLabels:
+      app: d
+  mtls:
+    mode: PERMISSIVE
+---
