+++ tests/integration/security/util/framework.go
@@ -45,6 +45,9 @@ func EchoConfig(name string, ns namespace.Instance, headless bool, annos echo.An
 				Protocol: protocol.GRPC,
 			},
 		},
+		WorkloadOnlyPorts: []int{
+			8085,
+		},
 		Galley: g,
 		Pilot:  p,
 	}
